Improve its performance and readability
