+++ code/app.py
@@ -2,8 +2,13 @@ import chainlit as cl
 from chainlit import Message
 from fastapi import Request, Response
 from db import save_interaction
-from rag_pipeline import process_message_with_memory
-from conversation_memory import session_memory, MessageRole
+from rag_pipeline import create_rag_chain
+from conversation_memory import (
+    session_memory, 
+    MessageRole,
+    create_conversation_context
+)
+from language_detection import detect_language_and_get_name
 from rss_reader import get_rss_items
 from custom_data_layer import CustomDataLayer
 from free_seats import get_occupancy_data, make_plotly_figure
@@ -12,24 +17,34 @@ from free_seats import get_occupancy_data, make_plotly_figure
 
 # === Authentication (optional) ===
 users = [
-    cl.User(identifier="1", display_name="Admin", metadata={"username": "admin", "password": "admin"})
+    cl.User(identifier="1", display_name="Admin", 
+            metadata={"username": "admin", "password": "admin"})
 ]
+
 
 # === Data Layer ===
 @cl.data_layer
 def get_data_layer():
     return CustomDataLayer()
 
+
 # === Starter Buttons ===
 @cl.set_starters
 async def set_starters():
     return [
-        cl.Starter(label="Öffnungszeiten", message="Ist die Bibliothek jetzt geöffnet?"),
-        cl.Starter(label="Sitzplätze", message="Gibt es aktuell freie Sitzplätze in der Bibliothek?"),
-        cl.Starter(label="Services", message="Liste alle Dienstleistungen an der UB Mannheim auf"),
+        cl.Starter(
+            label="Öffnungszeiten", 
+            message="Ist die Bibliothek jetzt geöffnet?"),
+        cl.Starter(
+            label="Sitzplätze", 
+            message="Gibt es aktuell freie Sitzplätze in der Bibliothek?"),
+        cl.Starter(
+            label="Services", 
+            message="Liste alle Dienstleistungen an der UB Mannheim auf"),
         cl.Starter(label="Standorte", message="Standorte der UB Mannheim"),
         cl.Starter(label="Neuigkeiten", message="Neues aus der UB"),
     ]
+
 
 # === Chat Start: Initialize Session Memory ===
 @cl.on_chat_start
@@ -37,14 +52,20 @@ async def on_chat_start():
     session_id = cl.user_session.get("id") or "unknown"
     cl.user_session.set("session_id", session_id)
     
+    # Create and store the RAG chain in the user session
+    rag_chain = await create_rag_chain()
+    cl.user_session.set("rag_chain", rag_chain)
+
     # Clear any existing session memory for this user
     session_memory.clear_session(session_id)
+
 
 # === Chat Message Handler ===
 @cl.on_message
 async def on_message(message: cl.Message):
     user_input = message.content.strip()
     session_id = cl.user_session.get("session_id") or "unknown"
+    rag_chain = cl.user_session.get("rag_chain")
 
     # RSS feed
     news_keywords = ["news", "neues", "neuigkeiten", "aktuelles", "nachrichten"]
@@ -96,9 +117,35 @@ async def on_message(message: cl.Message):
 
     # RAG Response with Memory
     try:
-        response = await process_message_with_memory(session_id, user_input)
-        await Message(content=response).send()
-        await save_interaction(session_id, user_input, response)
+        # Add user message to memory
+        session_memory.add_turn(session_id, MessageRole.USER, user_input)
+
+        # Detect language
+        detected_language = detect_language_and_get_name(user_input)
+
+        # Get conversation context
+        conversation_context = create_conversation_context(session_id)
+        
+        # Invoke RAG chain
+        response_generator = rag_chain.astream({
+            "question": user_input,
+            "conversation_context": conversation_context,
+            "language": detected_language
+        })
+
+        # Stream response
+        full_response = ""
+        msg = Message(content="", author="assistant")
+        async for token in response_generator:
+            await msg.stream_token(token)
+            full_response += token
+        await msg.send()
+        
+        # Add assistant response to memory
+        session_memory.add_turn(
+            session_id, MessageRole.ASSISTANT, full_response)
+        await save_interaction(session_id, user_input, full_response)
+
     except Exception as e:
         error_response = f"❌ Fehler bei der Verarbeitung: {str(e)}"
         await Message(content=error_response).send()
@@ -108,10 +155,15 @@ async def on_message(message: cl.Message):
         session_memory.add_turn(session_id, MessageRole.ASSISTANT, error_response)
         await save_interaction(session_id, user_input, error_response)
 
+
     # Optional: fallback to web search
     # fallback = search_ub_website(user_input)
-    # await Message(content=f"Ich konnte nichts Genaues finden. Ergebnisse von der UB-Website:\n\n{fallback}").send()
+    # await Message(
+    #     content=f"Ich konnte nichts Genaues finden. "
+    #             f"Ergebnisse von der UB-Website:\n\n{fallback}"
+    # ).send()
     # await save_interaction(session_id, user_input, fallback)
+
 
 # === Chat End ===
 @cl.on_chat_end
@@ -122,11 +174,13 @@ async def on_chat_end():
         session_memory.end_session(session_id)
     print("User disconnected.")
 
+
 # === Logout ===
 @cl.on_logout
 def on_logout(request: Request, response: Response):
     for cookie_name in request.cookies.keys():
         response.delete_cookie(cookie_name)
+
 
 # === Optional: Password Auth ===
 # @cl.password_auth_callback
